[CHORE] caret 아이콘 폴더 구조 변경 | Change caret folder TALEARNT-473
